update: 完成user-service

+++ common/src/main/java/com/altaria/common/pojos/user/entity/User.java
@@ -1,6 +1,8 @@
 package com.altaria.common.pojos.user.entity;
 
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -11,13 +13,13 @@ import java.time.LocalDateTime;
 @Setter
 @Getter
 public class User implements Serializable {
-    private Integer id;
+    private Long id;
     private String userName;
     private String password;
     private String email;
     private String nickName;
     private String avatar;
-    private int role;
+    private Integer role;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
 }
